+++ tests/test_parsing.py
@@ -2,10 +2,7 @@ from kg_maintainer.models import (
     CharacterProfile,
     WorldItem,
 )
-from kg_maintainer.parsing import (  # Assuming parsing is now a module inside kg_maintainer
-    parse_unified_character_updates,
-    parse_unified_world_updates,
-)
+from parsing import parse_unified_character_updates, parse_unified_world_updates
 
 # Example of a helper to create JSON strings for tests if they get complex
 # def make_json_string(data_dict):
