+++ tests/test_parsing_utils.py
@@ -2,7 +2,7 @@ import logging
 import sys
 import unittest
 
-from parsing_utils import parse_rdf_triples_with_rdflib
+from parsing import parse_rdf_triples_with_rdflib
 
 
 class TestRdfTripleParsing(unittest.TestCase):
