+++ utils/text_utils.py
@@ -1,6 +1,9 @@
+# utils/text_utils.py
+"""Text-related helper functions."""
+
 from config import settings
 
 
 def _is_fill_in(value: object) -> bool:
-    """Return True if ``value`` equals the fill-in placeholder."""
+    """Return ``True`` if ``value`` equals the fill-in placeholder."""
     return isinstance(value, str) and value.strip() == settings.FILL_IN
